# docs/swagger.yaml
definitions:
...
  form.FieldFormat:
    properties:
      id:
        type: string
      layers:
        items:
          additionalProperties: true
          type: object
        type: array
    type: object
  form.FieldTableColumn:
    properties:
      datasource_column_id:
        description: 表字段id
        type: integer
      datasource_column_relation_id:
        description: 表字段关联表id,0表示本数据表字段
        type: integer
      field_id:
        description: 控件id
        type: integer
      form_id:
        description: 表单id
        type: integer
      id:
        type: integer
      is_condition:
        description: 是否作为筛选条件 0-关闭 1-开启
        type: boolean
      show_name:
        description: 表头显示名称
        type: string
    type: object
...
  server.ApiResponse:
    properties:
      code:
        description: 状态码
        type: integer
      message:
        description: 状态短语
        type: string
      result:
        description: 数据结果集
        type: object
    type: object
...
  server.FieldAttrItem:
    properties:
      config:
        additionalProperties: true
        type: object
      id:
        description: 控件id
        type: integer
      key:
        description: 控件key
        type: string
      title:
        description: 控件名称
        type: string
      type:
        description: 控件类型
        type: string
    type: object
...
  server.FlowButtons:
    properties:
      button_name:
        description: 按钮名称
        type: string
      form_id:
        description: 表单id
        type: integer
      form_name:
        description: 表单名称
        type: string
    type: object
...
  server.FormArg:
    properties:
      description:
        description: 表单描述
        type: string
      footer:
        $ref: '#/definitions/form.FieldFormat'
        description: 底部配置
      name:
        description: 表单名称
        type: string
      page_type:
        description: 页面类型 1-表单 2-展示页
        type: integer
      plugins:
        description: 控件信息
        items:
          $ref: '#/definitions/form.FieldFormat'
        type: array
      source_table_id:
        description: 数据表id
        type: integer
      status:
        description: 表单状态
        type: boolean
      type:
        description: 表单类型 1-标准表单 2-弹窗
        type: integer
    required:
    - name
    - page_type
    - type
    type: object
...
  server.FormEditArg:
    properties:
      description:
        description: 表单描述
        type: string
      footer:
        $ref: '#/definitions/form.FieldFormat'
        description: 底部配置
      id:
        type: integer
      name:
        description: 表单名称
        type: string
      page_type:
        description: 页面类型 1-表单 2-展示页
        type: integer
      plugins:
        description: 控件信息
        items:
          $ref: '#/definitions/form.FieldFormat'
        type: array
      source_table_id:
        description: 数据表id
        type: integer
      status:
        description: 表单状态
        type: boolean
      type:
        description: 表单类型 1-标准表单 2-弹窗
        type: integer
    required:
    - id
    - name
    - page_type
    - type
    type: object
...
  server.FormItemVO:
    properties:
      app_id:
        description: 应用id
        type: integer
      content:
        description: 表单控件json
        items:
          type: integer
        type: array
      created_at:
        type: string
      datasource_table_id:
        description: 数据表id
        type: integer
      desc:
        description: 表单描述
        type: string
      footer:
        $ref: '#/definitions/form.FieldFormat'
      from:
        description: 原表
        type: integer
      id:
        type: integer
      is_delete:
        description: 删除状态 0-未删除 1-已删除
        type: boolean
      is_online:
        description: 是否是上线表单 1-是
        type: boolean
      name:
        description: 表单名称
        type: string
      number:
        description: 表单编号
        type: string
      page_type:
        description: 页面类型 1-表单 2-展示页
        type: integer
      plugins:
        items:
          $ref: '#/definitions/form.FieldFormat'
        type: array
      status:
        description: 表单状态 0-未生效 1-已生效
        type: boolean
      type:
        description: 表单类型 1-标准表单
        type: integer
      updated_at:
        type: string
    type: object
  server.FormListItem:
    properties:
      id:
        type: integer
      is_bound:
        description: 是否有流程依赖
        type: boolean
      name:
        description: 表单名称
        type: string
      "no":
        description: 表单编号
        type: string
      page_type:
        description: 页面类型
        type: integer
      status:
        description: 表单状态
        type: boolean
      type:
        description: 表单类型
        type: integer
      update_time:
        type: string
    type: object
  server.FormListVO:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/server.FormListItem'
        type: array
    type: object
...
  server.IdStruct:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  server.RelatedForm:
    properties:
      field_name:
        description: 控件名称
        type: string
      fields:
        description: 字段集合
        items:
          $ref: '#/definitions/server.FieldAttrItem'
        type: array
      form_id:
        description: 表单id
        type: integer
      name:
        description: 表单名称
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /form/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: id
        in: body
        name: IdStruct
        required: true
        schema:
          $ref: '#/definitions/server.IdStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  type: object
              type: object
      summary: 表单删除[new]
      tags:
      - form
  /form/detail:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 表单id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/server.FormItemVO'
              type: object
      summary: 表单详情
      tags:
      - form
  /form/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 请求体
        in: body
        name: FormEditArg
        required: true
        schema:
          $ref: '#/definitions/server.FormEditArg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  type: object
              type: object
      summary: 编辑表单
      tags:
      - form
  /form/fields:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 表单id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/server.FieldAttrItem'
                  type: array
              type: object
      summary: 表单字段列表
      tags:
      - form
  /form/flow/button:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 流程id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/server.FlowButtons'
                  type: array
              type: object
      summary: 绑定流程的按钮列表[new]
      tags:
      - form
  /form/list:
    get:
      consumes:
      - application/json
      description: 页码不传获取的为生效表单 传了页码则可分页获取全部表单
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: string
      - description: 页容量
        in: query
        name: limit
        type: string
      - description: 表单分类 1-标准表单 2-弹窗
        in: query
        name: type
        type: string
      - description: 表单分类 1-表单 2-显示
        in: query
        name: page_type
        type: string
      - description: 表单名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/server.FormListVO'
              type: object
      summary: 表单列表[update]
      tags:
      - form
  /form/list/collect:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 是否获取生效表单,否时不传或传空值,是时传递有效字符串即可
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/server.FormItemVO'
                  type: array
              type: object
      summary: 获取表单页表单
      tags:
      - form
...
  /form/new:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 请求体
        in: body
        name: FormArg
        required: true
        schema:
          $ref: '#/definitions/server.FormArg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  type: object
              type: object
      summary: 新增表单[update]
      tags:
      - form
  /form/relation:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 表单id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/server.RelatedForm'
                  type: array
              type: object
      summary: 获取关联列表
      tags:
      - field
  /form/table/columns:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 列表控件id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/form.FieldTableColumn'
                  type: array
              type: object
      summary: 列表控件字段列表[new]
      tags:
      - form
...
swagger: "2.0"